chore(tab): reorganize files
