+++ projects/cashmere/src/lib/tabs/index.ts
@@ -1,4 +1,4 @@
-export {TabComponent} from './tab.component';
-export {TabSetComponent, TabChangeEvent} from './tab-set.component';
-export {HcTabTitleComponent} from './tab-title.component';
+export {TabComponent} from './tab/tab.component';
+export {TabSetComponent, TabChangeEvent} from './tab-set/tab-set.component';
+export {HcTabTitleComponent} from './tab/tab-title.component';
 export {TabsModule} from './tabs.module';
