+++ projects/cashmere/src/lib/tabs/tab-set/tab-set.component.scss
@@ -1,4 +1,4 @@
-@import '../sass/colors';
+@import '../../sass/colors';
 
 .hc-horizontal-tab-container {
     height: auto;
