+++ projects/cashmere/src/lib/tabs/tab/tab.component.scss
@@ -1,4 +1,4 @@
-@import '../sass/colors';
+@import '../../sass/colors';
 
 :host {
     min-width: 0;
