+++ projects/cashmere/src/lib/tabs/tab/tab.component.spec.ts
@@ -1,7 +1,7 @@
 import {TabComponent} from './tab.component';
 import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 import {RouterTestingModule} from '@angular/router/testing';
-import { TabsService } from './tabs.service';
+import { TabsService } from '../tabs.service';
 
 describe('TabComponent', () => {
