+++ projects/cashmere/src/lib/tabs/tabs.module.ts
@@ -1,9 +1,9 @@
 import {RouterModule} from '@angular/router';
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {TabComponent} from './tab.component';
-import {HcTabTitleComponent} from './tab-title.component';
-import {TabSetComponent} from './tab-set.component';
+import {TabComponent} from './tab/tab.component';
+import {HcTabTitleComponent} from './tab/tab-title.component';
+import {TabSetComponent} from './tab-set/tab-set.component';
 
 @NgModule({
     imports: [CommonModule, RouterModule],
